affile: Error out when there's a syntax error in the filename template
When there is a syntax error in the filename template (eg,
"/var/log/foo-$(+ 0"), error out at config load time. This is done by
adding an error output variable to affile_dd_new_instance() and both
affile_dd_new() and afpipe_dd_new(), then using that to store the result
of log_template_compile(). If that function fails, we bail out early,
and use CHECK_ERROR() in the grammar to notify the user of the issue.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>