afmongodb: Implement bulk insert mode
This implements support for inserting messages in bulk, by adding a
flush-lines() option to the grammar, and constructing that many BSONs
and storing them internally, before inserting them in bulk. This
improves performance by a large margin.

WIP: We ACK the messages after formatting the BSON, instead of at insert
time, and there is no flush at deinit time. The code could use some
refactoring too, but this will serve as a reasonable PoC.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>