debian: Only use dpkg-maintscript-helper if available
If dpkg-maintscript-helper is not available, don't try to use it, but
leave the old conffiles there instead. They don't do any harm, and
adding code to the maintainer scripts to handle the case of
dpkg-maintscript-helper being unavailable is more trouble than it is
worth. So just leave the files there in that case.

Reported-by: Imre Lazar <imre@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>