csvparser: escape related flags are incorrectly processed
The problem is that escape related flags just simply overwrite
default flags and overwritten by non-escape flags.
This patch is do the following in order to avoid undesired owerwrites:

If new flag is an escape flag (and does not contain any other flags),
previously set flags, without escape flags, are appended to that new flag.

If new flag does not contain escape flags, escape flags from previously
set flags are appended to that new flag.

Otherwise new flag is overwrite previously set flags.

Please note, that this solution is not the best( adding a new escape-mode()
argument would be better than this workaround ).

Signed-off-by: Budai Laszlo <lbudai@balabit.hu>